rename message handler file
