+++ lib/TestScriptRunnable.rb
@@ -4,7 +4,7 @@ require 'jsonpath'
 require 'fhir_client'
 require_relative 'assertions'
 require_relative './TestReportHandler.rb'
-require_relative './MessageHandler.rb'
+require_relative 'message_handler'
 require_relative './Operation.rb'
 
 class TestScriptRunnable
