+++ lib/message_handler.rb
